+++ tests/conftest.py
@@ -9,6 +9,7 @@ import numpy as np
 from typing import Generator
 
 from pulser_simulation import QutipBackendV2
+from pulser.devices._device_datacls import BaseDevice
 from emu_sv import SVBackend
 from emu_mps import MPSBackend
 
@@ -309,3 +310,15 @@ def restore_rng_state() -> Generator:
 @pytest.fixture
 def make_mock_connection() -> type[MockConnection]:
     return MockConnection
+
+
+@pytest.fixture
+def device() -> BaseDevice:
+    return AnalogDeviceWithDMM()._device
+
+
+@pytest.fixture
+def max_min_dist_ratio(device: BaseDevice) -> float:
+    assert isinstance(device.max_radial_distance, int)
+    assert isinstance(device.min_atom_distance, int)
+    return device.max_radial_distance / device.min_atom_distance
