+++ tests/solvers/test_decomposition.py
@@ -83,6 +83,7 @@ def test_initial_steps_solver(decomposable_qubo: QUBOInstance, use_quantum: bool
     first_vertex = 0
 
     pulser_device = solver._solver.device._pulser_device
+    assert pulser_device.max_radial_distance is not None
     min_distance, max_radial_distance = compute_min_max_distances(
         qubo_mat,
         max_min_dist_ratio=pulser_device.max_radial_distance / pulser_device.min_atom_distance,
